large refactoring and moving without change of functionality
