+++ pdf_generator/templates/default/base64_decoder.tex
