+++ pdf_generator/templates/default/binwalk.tex
