+++ pdf_generator/templates/default/cpu_architecture.tex
