+++ pdf_generator/templates/default/exploit_mitigations.tex
