+++ pdf_generator/templates/default/init_systems.tex
