+++ pdf_generator/templates/default/printable_strings.tex
