+++ pdf_generator/templates/default/string_evaluator.tex
