+++ pdf_generator/templates/fact_logo.png
