--- src/config/main.cfg
@@ -1,4 +1,0 @@
-
-[Logging]
-logFile=/tmp/fact_report.log
-logLevel=WARNING
