+++ test/acceptance/test_main.py
@@ -1,9 +1,9 @@
-from common_helper_process.fail_safe_subprocess import execute_shell_command_get_return_code
-import os
+from pathlib import Path
+
 import pytest
+from common_helper_process.fail_safe_subprocess import execute_shell_command_get_return_code
 
-
-SRC_DIR = os.path.dirname(os.path.abspath(__file__)) + '/../../pdf_generator.py'
+SCRIPT_PATH = Path(__file__).parent.parent.parent / 'pdf_generator' / 'pdf_generator.py'
 
 
 @pytest.mark.parametrize('arguments, expected_output, expected_return_code', [
@@ -11,7 +11,7 @@ SRC_DIR = os.path.dirname(os.path.abspath(__file__)) + '/../../pdf_generator.py'
     ('-h', 'usag', 0)
 ])
 def test_main_program(arguments, expected_output, expected_return_code):
-    command_line = SRC_DIR + ' ' + arguments
+    command_line = str(SCRIPT_PATH) + ' ' + arguments
     output, return_code = execute_shell_command_get_return_code(command_line)
     assert output[0:4] == expected_output
     assert return_code == expected_return_code
